Xext: xf86bigfont: move it into it's own subdir
