+++ meson.build
@@ -841,7 +841,8 @@ if build_xserver
         libxserver_record,
         libxserver_os,
         libxserver_dri2,
-        libxserver_dri3
+        libxserver_dri3,
+        libxserver_xf86bigfont
     ]
 
     if build_namespace
